requirements.txt
datasets
sacrebleu
torchtext
flask
accelerate
underthesea
datasets